feat: 가게 소유자 확인 로직 추가 #52

+++ src/main/java/com/market/product/service/ProductService.java
@@ -1,11 +1,15 @@
 package com.market.product.service;
 
 import com.market.core.code.error.MarketErrorCode;
+import com.market.core.code.error.MemberErrorCode;
 import com.market.core.code.error.ProductErrorCode;
 import com.market.core.exception.MarketException;
+import com.market.core.exception.MemberException;
 import com.market.core.exception.ProductException;
 import com.market.market.entity.Market;
 import com.market.market.repository.MarketRepository;
+import com.market.member.entity.Member;
+import com.market.member.repository.MemberRepository;
 import com.market.product.dto.request.ProductCreateRequest;
 import com.market.product.dto.request.ProductUpdateRequest;
 import com.market.product.dto.response.ProductResponse;
@@ -25,14 +29,18 @@ public class ProductService {
 
     private final ProductRepository productRepository;
     private final MarketRepository marketRepository;
+    private final MemberRepository memberRepository;
 
     /**
      * 상품을 등록합니다.
      */
     @Transactional
-    public void createProduct(Long marketId, ProductCreateRequest productCreateRequest) {
+    public void createProduct(Long memberId, Long marketId, ProductCreateRequest productCreateRequest) {
         Market market = marketRepository.findById(marketId)
                 .orElseThrow(() -> new MarketException(MarketErrorCode.NOT_FOUND_MARKET_ID));
+
+        // 가게 소유자가 맞는지 확인
+        verifyOwnerOfMarket(memberId, marketId);
 
         Product product = Product.builder()
                 .market(market)
@@ -73,9 +81,12 @@ public class ProductService {
      * 상품을 수정합니다.
      */
     @Transactional
-    public void updateProduct(Long productId, ProductUpdateRequest productUpdateRequest) {
+    public void updateProduct(Long memberId, Long productId, ProductUpdateRequest productUpdateRequest) {
         Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new ProductException(ProductErrorCode.NOT_FOUND_PRODUCT_ID));
+
+        // 가게 소유자가 맞는지 확인
+        verifyOwnerOfMarket(memberId, product.getMarket().getId());
 
         product.updateProduct(productUpdateRequest);
     }
@@ -84,10 +95,28 @@ public class ProductService {
      * 상품을 삭제합니다.
      */
     @Transactional
-    public void deleteProduct(Long productId) {
+    public void deleteProduct(Long memberId, Long productId) {
         Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new ProductException(ProductErrorCode.NOT_FOUND_PRODUCT_ID));
 
+        // 가게 소유자가 맞는지 확인
+        verifyOwnerOfMarket(memberId, product.getMarket().getId());
+
         productRepository.delete(product);
+    }
+
+    /**
+     * 가게 소유지가 맞는지 확인합니다.
+     */
+    private void verifyOwnerOfMarket(Long memberId, Long marketId) {
+        Market market = marketRepository.findById(marketId)
+                .orElseThrow(() -> new MarketException(MarketErrorCode.NOT_FOUND_MARKET_ID));
+
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new MemberException(MemberErrorCode.NOT_FOUND_MEMBER_ID));
+
+        if (!market.getMember().equals(member)) {
+            throw new MarketException(MarketErrorCode.NOT_OWN_STORE);
+        }
     }
 }
